Move all validation to its own module.
