+++ tests/ludwig/config_validation/test_validate_config_combiner.py
@@ -1,8 +1,8 @@
 import pytest
 from jsonschema.exceptions import ValidationError
 
+from ludwig.config_validation.validation import validate_config
 from ludwig.constants import TRAINER
-from ludwig.schema import validate_config
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature
