+++ tests/ludwig/config_validation/test_validate_config_features.py
@@ -1,7 +1,7 @@
 import pytest
 from jsonschema.exceptions import ValidationError
 
-from ludwig.schema import validate_config
+from ludwig.config_validation.validation import validate_config
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature, text_feature
