+++ tests/ludwig/config_validation/test_validate_config_trainer.py
@@ -1,8 +1,8 @@
 import pytest
 from jsonschema.exceptions import ValidationError
 
+from ludwig.config_validation.validation import validate_config
 from ludwig.constants import TRAINER
-from ludwig.schema import validate_config
 from ludwig.schema.optimizers import optimizer_registry
 from ludwig.schema.trainer import ECDTrainerConfig
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature
